+++ skyux-spa-visual-tests/src/app/definition-list/definition-list.visual-spec.ts
@@ -1,0 +1,27 @@
+import { SkyVisualTest } from '../../../config/utils/visual-test-commands';
+
+describe('Definition list', () => {
+
+  it('should match definition list screenshot with default settings', () => {
+    return SkyVisualTest.setupTest('definition-list')
+    .then(() => {
+      return SkyVisualTest.compareScreenshot({
+        screenshotName: 'definition-list-defaults',
+        selector: '#screenshot-definition-list-1'
+      });
+    });
+
+  });
+
+  it('should match definition list screenshot with custom settings', () => {
+    return SkyVisualTest.setupTest('definition-list')
+    .then(() => {
+      return SkyVisualTest.compareScreenshot({
+        screenshotName: 'definition-list-overrides',
+        selector: '#screenshot-definition-list-2'
+      });
+    });
+
+  });
+
+});