+++ skyux-spa-visual-tests/src/app/navbar/navbar-visual.component.ts
@@ -1,0 +1,7 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'navbar-visual',
+  templateUrl: './navbar-visual.component.html'
+})
+export class NavbarVisualComponent { }