+++ src/modules/error/error-modal-config.ts
@@ -1,3 +1,4 @@
+/* istanbul ignore next */
 export class ErrorModalConfig {
   public errorTitle: string;
   public errorDescription: string;