+++ src/modules/list-filters/list-filter-inline.component.ts
@@ -26,11 +26,10 @@ import {
   templateUrl: './list-filter-inline.component.html'
 })
 export class SkyListFilterInlineComponent implements AfterContentInit {
+  public inlineFilters: Array<SkyListFilterInlineModel> = [];
 
   @ContentChildren(SkyListFilterInlineItemComponent)
   private filters: QueryList<SkyListFilterInlineItemComponent>;
-
-  private inlineFilters: Array<SkyListFilterInlineModel> = [];
 
   constructor(
     private ref: ChangeDetectorRef,
@@ -72,5 +71,4 @@ export class SkyListFilterInlineComponent implements AfterContentInit {
       });
     });
   }
-
 }