+++ src/modules/list-toolbar/list-toolbar.component.ts
@@ -28,7 +28,8 @@ import {
   ListStateDispatcher,
   ListSortLabelModel,
   ListSortFieldSelectorModel,
-  ListFilterModel
+  ListFilterModel,
+  ListPagingSetPageNumberAction
 } from '../list/state';
 
 import { SkyListToolbarItemComponent } from './list-toolbar-item.component';
@@ -268,12 +269,20 @@ export class SkyListToolbarComponent implements OnInit, AfterContentInit {
     this.inlineFilterBarExpanded = !this.inlineFilterBarExpanded;
   }
 
-  private itemIsInView(itemView: string, activeView: string) {
-    return (itemView === undefined || itemView === activeView);
+  public updateSearchText(searchText: string) {
+    this.state.take(1).subscribe((currentState) => {
+      this.dispatcher.searchSetText(searchText);
+      if (currentState.paging.pageNumber && currentState.paging.pageNumber !== 1) {
+        this.dispatcher.next(
+          new ListPagingSetPageNumberAction(Number(1))
+        );
+      }
+    });
+
   }
 
-  private updateSearchText(searchText: string) {
-    this.dispatcher.searchSetText(searchText);
+  private itemIsInView(itemView: string, activeView: string) {
+    return (itemView === undefined || itemView === activeView);
   }
 
   private getSortSelectors() {