+++ src/modules/list-view-checklist/list-view-checklist.component.ts
@@ -189,6 +189,8 @@ export class SkyListViewChecklistComponent extends ListViewComponent implements 
           return true;
         }
       }
+
+      return false;
     };
   }