+++ src/modules/list/state/toolbar/toolbar.orchestrator.ts
@@ -10,7 +10,7 @@ import {
 
 export class ListToolbarOrchestrator
   extends ListStateOrchestrator<ListToolbarModel> {
-
+  /* istanbul ignore next */
   constructor() {
     super();