+++ src/modules/repeater/repeater.module.ts
@@ -1,6 +1,8 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+
 import { SkyChevronModule } from '../chevron';
 import { SkyLogModule } from '../log';
 
@@ -10,6 +12,8 @@ import { SkyRepeaterItemTitleComponent } from './repeater-item-title.component';
 import { SkyRepeaterItemComponent } from './repeater-item.component';
 import { SkyRepeaterComponent } from './repeater.component';
 import { SkyRepeaterService } from './repeater.service';
+import { SkyCheckboxModule } from '../checkbox';
+import { SkyResourcesModule } from '../resources/resources.module';
 
 @NgModule({
   declarations: [
@@ -23,9 +27,12 @@ import { SkyRepeaterService } from './repeater.service';
     SkyRepeaterService
   ],
   imports: [
+    BrowserAnimationsModule,
     CommonModule,
     SkyChevronModule,
-    SkyLogModule
+    SkyLogModule,
+    SkyCheckboxModule,
+    SkyResourcesModule
   ],
   exports: [
     SkyRepeaterComponent,