+++ src/modules/text-expand-repeater/text-expand-repeater.component.ts
@@ -28,11 +28,12 @@ export class SkyTextExpandRepeaterComponent implements AfterViewInit {
     this.setup(value);
   }
   public buttonText: string;
-  private _data: Array<any>;
+  public contentItems: Array<any>;
+  public expandable: boolean;
+
   private seeMoreText: string = this.resources.getString('text_expand_see_more');
   private seeLessText: string = this.resources.getString('text_expand_see_less');
   private isExpanded: boolean = false;
-  private expandable: boolean;
   @ViewChild('container')
   private containerEl: ElementRef;
   private items: Array<HTMLElement>;
@@ -41,9 +42,9 @@ export class SkyTextExpandRepeaterComponent implements AfterViewInit {
     private textExpandRepeaterAdapter: SkyTextExpandRepeaterAdapterService) { }
 
   public ngAfterViewInit() {
-    if (this._data) {
+    if (this.contentItems) {
       this.items = this.textExpandRepeaterAdapter.getItems(this.elRef);
-      for (let i = this.maxItems; i < this._data.length; i++) {
+      for (let i = this.maxItems; i < this.contentItems.length; i++) {
         this.textExpandRepeaterAdapter.hideItem(this.items[i]);
       }
     }
@@ -52,27 +53,56 @@ export class SkyTextExpandRepeaterComponent implements AfterViewInit {
   public animationEnd() {
     // Ensure the correct items are displayed
     if (!this.isExpanded) {
-      for (let i = this.maxItems; i < this._data.length; i++) {
+      for (let i = this.maxItems; i < this.contentItems.length; i++) {
         this.textExpandRepeaterAdapter.hideItem(this.items[i]);
       }
     }
     // Set height back to auto so the browser can change the height as needed with window changes
-    this.textExpandRepeaterAdapter.setContainerHeight(this.containerEl, 'auto');
+    this.textExpandRepeaterAdapter.setContainerHeight(this.containerEl, undefined);
   }
 
   public repeaterExpand() {
+    if (!this.isExpanded) {
+      this.setContainerMaxHeight();
+      setTimeout(() => {
+        this.isExpanded = true;
+        this.animateRepeater(true);
+      });
+
+    } else {
+      this.setContainerMaxHeight();
+      setTimeout(() => {
+        this.isExpanded = false;
+        this.animateRepeater(false);
+      });
+
+    }
+  }
+
+  private setContainerMaxHeight() {
+    // ensure everything is reset
+    this.animationEnd();
+    /* Before animation is kicked off, ensure that a maxHeight exists */
+    /* Once we have support for angular v4 animations with parameters we can use that instead */
+    let currentHeight = this.textExpandRepeaterAdapter.getContainerHeight(this.containerEl);
+    this.textExpandRepeaterAdapter.setContainerHeight(this.containerEl, `${currentHeight}px`);
+  }
+
+  private animateRepeater(expanding: boolean) {
     let adapter = this.textExpandRepeaterAdapter;
     let container = this.containerEl;
+
+    adapter.setContainerHeight(container, undefined);
     let currentHeight = adapter.getContainerHeight(container);
-    for (let i = this.maxItems; i < this._data.length; i++) {
-      if (this.isExpanded) {
+    for (let i = this.maxItems; i < this.contentItems.length; i++) {
+      if (!expanding) {
         adapter.hideItem(this.items[i]);
       } else {
         adapter.showItem(this.items[i]);
       }
     }
     let newHeight = adapter.getContainerHeight(container);
-    if (this.isExpanded) {
+    if (!expanding) {
       this.buttonText = this.seeMoreText;
     } else {
       this.buttonText = this.seeLessText;
@@ -80,17 +110,21 @@ export class SkyTextExpandRepeaterComponent implements AfterViewInit {
     if (newHeight < currentHeight) {
       // The new text is smaller than the old text, so put the old text back before doing
       // the collapse animation to avoid showing a big chunk of whitespace.
-      for (let i = this.maxItems; i < this._data.length; i++) {
+      for (let i = this.maxItems; i < this.contentItems.length; i++) {
         adapter.showItem(this.items[i]);
       }
     }
+
     adapter.setContainerHeight(container, `${currentHeight}px`);
     // This timeout is necessary due to the browser needing to pick up the non-auto height being set
     // in order to do the transtion in height correctly. Without it the transition does not fire.
-    setTimeout(function () {
+    setTimeout(() => {
       adapter.setContainerHeight(container, `${newHeight}px`);
-    }, 5);
-    this.isExpanded = !this.isExpanded;
+      /* This resets values if the transition does not get kicked off */
+      setTimeout(() => {
+        this.animationEnd();
+      }, 500);
+    }, 10);
   }
 
   private setup(value: Array<any>) {
@@ -103,9 +137,9 @@ export class SkyTextExpandRepeaterComponent implements AfterViewInit {
       } else {
         this.expandable = false;
       }
-      this._data = value;
+      this.contentItems = value;
     } else {
-      this._data = undefined;
+      this.contentItems = undefined;
       this.expandable = false;
     }
   }